// src/types/enums/dealer.ts
export enum Dealer {
  // UpperCase for BE matching
  BESSERGOLD_CZ = "BESSERGOLD_CZ",
  BESSERGOLD_DE = "BESSERGOLD_DE",
  SILVERUM = "SILVERUM",
  ZLATAKY = "ZLATAKY",
  CESKA_MINCOVNA = "CESKA_MINCOVNA",
  GOLD_A_SILVER = "GOLD_A_SILVER",
  AURUM_PRO = "AURUM_PRO",
  AURUM_BOHEMICA = "AURUM_BOHEMICA",
  EKKA_GOLD = "EKKA_GOLD",
  JEDNA_UNCE = "JEDNA_UNCE",
  GOLD_SAFE = "GOLD_SAFE",
  GOLDEN_HOUSE = "GOLDEN_HOUSE",
}